split composite interaction implementation from header files
